benchmarks/hard/pyproject.toml
[project]
name = "kernelbench-hard"
version = "0.1.0"
description = "Surgical GPU kernel benchmark. Small deck, frontier agents, roofline metric, link-don't-spoil problem briefs."
requires-python = ">=3.11"
dependencies = [
    "torch>=2.11",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "rich>=13.7",
    "matplotlib>=3.9",
    "pytest>=8.0",
    "ruff>=0.6",
    "einops>=0.8.2",
    "requests<2.33",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
